+++ R/PrepData.R
@@ -1,7 +1,7 @@
 PrepData <- function(data, model, itemtype, guess, upper, gpcm_mats,
                      parprior, verbose, technical, parnumber = 1, BFACTOR = FALSE,
                      grsm.block = NULL, rsm.block = NULL, mixed.design, customItems,
-                     customGroup, fulldata = NULL, key, internal_constraints)
+                     customGroup, fulldata = NULL, key, spline_args, internal_constraints)
 {
     if(is.null(grsm.block)) grsm.block <- rep(1, ncol(data))
     if(is.null(rsm.block)) rsm.block <- rep(1, ncol(data))
@@ -122,7 +122,7 @@ PrepData <- function(data, model, itemtype, guess, upper, gpcm_mats,
                            itemnames=itemnames, exploratory=exploratory, parprior=parprior,
                            parnumber=parnumber, BFACTOR=BFACTOR, mixed.design=mixed.design,
                            customItems=customItems, customGroup=customGroup, key=key,
-                           gpcm_mats=gpcm_mats)
+                           gpcm_mats=gpcm_mats, spline_args=spline_args)
     prodlist <- attr(pars, 'prodlist')
     exploratory <- attr(pars, 'exploratory')
     if(is(pars[[1L]], 'numeric') || is(pars[[1L]], 'logical')){