+++ R/mirt-package.R
@@ -26,7 +26,7 @@
 #' @title Full information maximum likelihood estimation of IRT models.
 #' @author Phil Chalmers \email{rphilip.chalmers@@gmail.com}
 #' @useDynLib mirt
-#' @import stats lattice GPArotation Rcpp stats4 methods sfsmisc mgcv
+#' @import stats lattice GPArotation Rcpp stats4 methods sfsmisc mgcv splines
 #' @importFrom utils write.table flush.console packageVersion
 #' @importFrom graphics symbols
 #' @exportMethod anova residuals summary logLik vcov