+++ NEWS.md
@@ -1,5 +1,10 @@
 # Changes in mirt 1.19
 
+- new `spline` itemtype to model B-spline response functions for dichotomous models. Useful for 
+  diagnostic purposes after detecting item-misfit. Additional arguments can be passed to the 
+  `spline_args` list input to control the behaviour of the splines for each item. Currently limited
+  to unidimensional models only
+
 - `fscores()` gains a `plausible.type` argument to select between normal approximation PVs or 
   Metropolis-Hastings samples (suggested by Yang Liu)